feat: better blueprint picker
