+++ .release-please-manifest.json
@@ -1,4 +1,4 @@
 {
   "packages/dm-core": "1.21.0",
-  "packages/dm-core-plugins": "1.26.0"
+  "packages/dm-core-plugins": "1.25.0"
 }
