+++ example/app/data/DemoDataSource/recipes/plugins/grid/car_grid/tyre.recipe.json
@@ -53,13 +53,14 @@
       "dimensions": "*",
       "config": {
         "type": "PLUGINS:dm-core-plugins/list/ListPluginConfig",
+        "viewConfigs": [],
+        "openAsTab": false,
         "headers": ["name", "compound"],
         "functionality": {
           "type": "PLUGINS:dm-core-plugins/list/FunctionalityConfig",
           "delete": true,
           "add": true,
-          "sort": true,
-          "expand": true
+          "sort": true
         }
       }
     }
