+++ example/app/data/DemoDataSource/recipes/plugins/grid/car_grid/tyreList.recipe.json
@@ -28,6 +28,7 @@
       "dimensions": "*",
       "config": {
         "type": "PLUGINS:dm-core-plugins/list/ListPluginConfig",
+        "viewConfigs": [],
         "headers": ["name", "compound"]
       }
     }
