+++ example/app/data/DemoDataSource/recipes/plugins/grid/example/order.recipe.json
@@ -28,6 +28,7 @@
       "dimensions": "*",
       "config": {
         "type": "PLUGINS:dm-core-plugins/list/ListPluginConfig",
+        "viewConfigs": [],
         "headers": ["name", "description"],
         "functionality": {
           "type": "PLUGINS:dm-core-plugins/list/FunctionalityConfig",
