+++ example/app/data/DemoDataSource/recipes/plugins/grid/example/orderItem.recipe.json
@@ -43,11 +43,12 @@
       "dimensions": "*",
       "config": {
         "type": "PLUGINS:dm-core-plugins/list/ListPluginConfig",
+        "viewConfigs": [],
         "headers": ["name", "quantity", "type"],
+        "openAsTab": true,
         "functionality": {
           "type": "PLUGINS:dm-core-plugins/list/FunctionalityConfig",
-          "delete": true,
-          "expand": true
+          "delete": true
         }
       }
     }
