+++ example/app/data/DemoDataSource/recipes/plugins/list/order_example/order.recipe.json
@@ -28,11 +28,12 @@
       "dimensions": "*",
       "config": {
         "type": "PLUGINS:dm-core-plugins/list/ListPluginConfig",
+        "viewConfigs": [],
         "headers": ["name", "description"],
+        "openAsTab": true,
         "functionality": {
           "type": "PLUGINS:dm-core-plugins/list/FunctionalityConfig",
-          "delete": true,
-          "expand": true
+          "delete": true
         }
       }
     }
