+++ example/app/data/DemoDataSource/recipes/plugins/list/order_example/orderItem.recipe.json
@@ -43,22 +43,13 @@
       "dimensions": "*",
       "config": {
         "type": "PLUGINS:dm-core-plugins/list/ListPluginConfig",
+        "viewConfigs": [],
         "headers": ["quantity", "type"],
+        "openAsTab": true,
         "functionality": {
           "type": "PLUGINS:dm-core-plugins/list/FunctionalityConfig",
           "delete": true,
-          "add": true,
-          "open": true,
-          "expand": true
-        },
-        "openViewConfig": {
-          "type": "CORE:InlineRecipeViewConfig",
-          "recipe": {
-            "name": "Yaml",
-            "type": "CORE:UiRecipe",
-            "plugin": "@development-framework/dm-core-plugins/yaml"
-          },
-          "scope": "self"
+          "add": true
         }
       }
     }
