+++ example/app/data/DemoDataSource/recipes/plugins/list/task_list/task.recipe.json
@@ -53,6 +53,8 @@
       "dimensions": "*",
       "config": {
         "type": "PLUGINS:dm-core-plugins/list/ListPluginConfig",
+        "viewConfigs": [],
+        "openAsTab": false,
         "selectFromScope": "dmss://DemoDataSource/plugins/list/task_list",
         "headers": ["name", "assigned"],
         "saveExpanded": false,
@@ -60,9 +62,7 @@
           "type": "PLUGINS:dm-core-plugins/list/FunctionalityConfig",
           "delete": true,
           "add": true,
-          "sort": true,
-          "expand": true,
-          "open": false
+          "sort": true
         }
       }
     }
