+++ example/app/data/DemoDataSource/recipes/plugins/list/task_list/taskList.recipe.json
@@ -28,6 +28,7 @@
       "dimensions": "*",
       "config": {
         "type": "PLUGINS:dm-core-plugins/list/ListPluginConfig",
+        "viewConfigs": [],
         "headers": ["name", "assigned"]
       }
     }
