+++ packages/dm-core-plugins/blueprints/list/FunctionalityConfig.json
@@ -34,20 +34,6 @@
       "attributeType": "boolean",
       "optional": true,
       "default": true
-    },
-    {
-      "name": "open",
-      "type": "CORE:BlueprintAttribute",
-      "attributeType": "boolean",
-      "optional": true,
-      "default": false
-    },
-    {
-      "name": "expand",
-      "type": "CORE:BlueprintAttribute",
-      "attributeType": "boolean",
-      "optional": true,
-      "default": true
     }
   ]
 }
