+++ packages/dm-core-plugins/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@development-framework/dm-core-plugins",
   "license": "MIT",
-  "version": "1.26.0",
+  "version": "1.25.0",
   "main": "dist/index.js",
   "dependencies": {
     "@development-framework/dm-core": "^1.21.0",
@@ -11,6 +11,7 @@
     "axios": "^1.4.0",
     "dompurify": "^3.0.6",
     "highlight.js": "^11.8.0",
+    "lodash": "^4.17.21",
     "luxon": "^3.4.3",
     "mermaid": "^10.0.0",
     "react-hook-form": "^7.48.2",
@@ -26,6 +27,7 @@
     "@types/dompurify": "^3.0.4",
     "@types/jest": "^29.5.0",
     "@types/js-yaml": "^4.0.5",
+    "@types/lodash": "^4.14.202",
     "@types/luxon": "^3.3.3",
     "@types/node": "^20.10.0",
     "@types/react-dom": "^18.2.7",
