+++ packages/dm-core-plugins/src/list/ListPlugin.tsx
@@ -22,32 +22,30 @@ import { AppendButton, ListItemButton, FormButton } from './Components'
 type TListConfig = {
   expanded?: boolean
   headers: string[]
-  expandViewConfig: TViewConfig
-  openViewConfig: TViewConfig
+  defaultView: TViewConfig
+  views: TViewConfig[]
+  openAsTab: boolean
   saveExpanded: boolean
   selectFromScope?: string
   functionality: {
     add: boolean
     sort: boolean
     delete: boolean
-    expand: boolean
-    open: boolean
   }
   resolveReferences: boolean
 }
 const defaultConfig: TListConfig = {
   expanded: false,
   headers: ['name', 'type'],
-  expandViewConfig: { type: 'ViewConfig', scope: 'self' },
-  openViewConfig: { type: 'ViewConfig' },
+  defaultView: { type: 'ViewConfig', scope: 'self' },
+  views: [],
+  openAsTab: false,
   saveExpanded: false,
   selectFromScope: undefined,
   functionality: {
     add: true,
     sort: true,
     delete: true,
-    expand: true,
-    open: false,
   },
   resolveReferences: true,
 }
@@ -59,6 +57,7 @@ export const ListPlugin = (props: IUIPlugin & { config?: TListConfig }) => {
     ...config,
     functionality: { ...defaultConfig.functionality, ...config.functionality },
   }
+
   const {
     items,
     attribute,
@@ -93,31 +92,32 @@ export const ListPlugin = (props: IUIPlugin & { config?: TListConfig }) => {
     updateItem(item, data, false)
   }
 
-  const handleExpand = (item: TItem<any>) => {
-    const isExpanded = expanded[item.key] || false
-    setExpanded({ ...expanded, [item.key]: !isExpanded })
-  }
-
-  const handleOpen = (item: TItem<any>) => {
-    if (!onOpen) {
-      toast.error(
-        'Invalid UiRecipes. The list plugin was not passed an "onOpen()"-function.'
+  function expandOrOpen(item: TItem<any>) {
+    if (internalConfig.openAsTab) {
+      const view = { label: item?.data?.name, type: 'ViewConfig' }
+      if (!onOpen) {
+        toast.error(
+          'Invalid UiRecipes. The list plugin was not passed an "onOpen()"-function.'
+        )
+        return
+      }
+      onOpen(
+        item.key,
+        view,
+        `${idReference}[${item.index}]`,
+        false,
+        config.saveExpanded
+          ? undefined
+          : (data: any) => handleItemUpdate(item, data),
+        config.saveExpanded
+          ? (data: any) => handleItemUpdate(item, data)
+          : undefined
       )
       return
+    } else {
+      const isExpanded = expanded[item.key] || false
+      setExpanded({ ...expanded, [item.key]: !isExpanded })
     }
-    const view = { ...internalConfig.openViewConfig, label: item?.data?.name }
-    onOpen(
-      item.key,
-      view,
-      `${idReference}[${item.index}]`,
-      false,
-      config.saveExpanded
-        ? undefined
-        : (data: any) => handleItemUpdate(item, data),
-      config.saveExpanded
-        ? (data: any) => handleItemUpdate(item, data)
-        : undefined
-    )
   }
 
   if (error) throw new Error(JSON.stringify(error, null, 2))
@@ -157,29 +157,37 @@ export const ListPlugin = (props: IUIPlugin & { config?: TListConfig }) => {
               }}
             >
               <Stack direction='row' spacing={0.5} alignItems='center'>
-                {internalConfig.functionality.expand && (
-                  <Tooltip title={expanded[item.key] ? 'Minimize' : 'Expand'}>
-                    <Button
-                      variant='ghost_icon'
-                      color='secondary'
-                      disabled={!item.isSaved}
-                      data-testid={`expandListItem-${index}`}
-                      onClick={() => handleExpand(item)}
-                    >
-                      <Icon
-                        data={chevron_right}
-                        size={24}
-                        title={expanded[item.key] ? 'Close item' : 'Open item'}
-                        className='transition-all'
-                        style={{
-                          transform: expanded[item.key]
-                            ? 'rotate(90deg)'
-                            : 'rotate(0deg)',
-                        }}
-                      />
-                    </Button>
-                  </Tooltip>
-                )}
+                <Tooltip
+                  title={
+                    !internalConfig.openAsTab
+                      ? expanded[item.key]
+                        ? 'Minimize'
+                        : 'Expand'
+                      : 'Open in new tab'
+                  }
+                >
+                  <Button
+                    variant='ghost_icon'
+                    color='secondary'
+                    disabled={!item.isSaved}
+                    data-testid={`expandListItem-${index}`}
+                    onClick={() => expandOrOpen(item)}
+                  >
+                    <Icon
+                      data={
+                        internalConfig.openAsTab ? external_link : chevron_right
+                      }
+                      size={internalConfig.openAsTab ? 18 : 24}
+                      title={expanded[item.key] ? 'Close item' : 'Open item'}
+                      className='transition-all'
+                      style={{
+                        transform: expanded[item.key]
+                          ? 'rotate(90deg)'
+                          : 'rotate(0deg)',
+                      }}
+                    />
+                  </Button>
+                </Tooltip>
                 {attribute && !attribute.contained && <Icon data={link} />}
                 {internalConfig.headers.map(
                   (attribute: string, index: number) => {
@@ -193,6 +201,7 @@ export const ListPlugin = (props: IUIPlugin & { config?: TListConfig }) => {
                           key={attribute}
                           variant='body_short'
                           bold={index === 0}
+                          onClick={() => expandOrOpen(item)}
                           style={{ cursor: 'pointer' }}
                         >
                           {item?.data[attribute]}
@@ -203,23 +212,6 @@ export const ListPlugin = (props: IUIPlugin & { config?: TListConfig }) => {
                 )}
               </Stack>
               <Stack direction='row' alignItems='center'>
-                {internalConfig.functionality.open && (
-                  <Tooltip title='Open in new tab'>
-                    <Button
-                      variant='ghost_icon'
-                      color='secondary'
-                      disabled={!item.isSaved}
-                      data-testid={`open-list-item-${index}`}
-                      onClick={() => handleOpen(item)}
-                    >
-                      <Icon
-                        data={external_link}
-                        size={18}
-                        title={'Open item'}
-                      />
-                    </Button>
-                  </Tooltip>
-                )}
                 {internalConfig.functionality.sort && (
                   <>
                     <ListItemButton
@@ -271,7 +263,10 @@ export const ListPlugin = (props: IUIPlugin & { config?: TListConfig }) => {
                         )
                       : `${idReference}[${item.index}]`
                   }
-                  viewConfig={internalConfig.expandViewConfig}
+                  viewConfig={
+                    internalConfig.views[item.index] ??
+                    internalConfig.defaultView
+                  }
                 />
               )}
             </Stack>
