+++ packages/dm-core/src/components/Pickers/Datepicker/Calendar.tsx
@@ -27,12 +27,8 @@ interface CalendarProps {
 export const Calendar = (props: CalendarProps): ReactElement => {
   const { dateTime, handleDateSelection } = props
   const [showMonthPicker, setShowMonthPicker] = useState(false)
-  const [activeMonth, setActiveMonth] = useState<number>(
-    dateTime.month ?? THIS_MONTH
-  )
-  const [activeYear, setActiveYear] = useState<number>(
-    dateTime.year ?? THIS_YEAR
-  )
+  const [activeMonth, setActiveMonth] = useState<number>(THIS_MONTH)
+  const [activeYear, setActiveYear] = useState<number>(THIS_YEAR)
   const cal = calendar(activeMonth, activeYear)
 
   const currentMonthName = Object.keys(CALENDAR_MONTHS)[activeMonth - 1]
