+++ packages/dm-core/src/components/Pickers/Datepicker/Datepicker.tsx
@@ -35,9 +35,7 @@ export const Datepicker = (props: DatepickerProps): ReactElement => {
   const [open, setOpen] = useState(false)
   const datepickerRef = useRef<any | null>(null)
   const [datetime, setDatetime] = useState(
-    selectedDate
-      ? DateTime.fromISO(selectedDate).toUTC()
-      : DateTime.now().toUTC()
+    DateTime.fromISO(selectedDate).toUTC()
   )
   const [fieldDateValue, setFieldDateValue] = useState(
     datetime.toFormat('dd/MM/yyyy') ?? 'dd/mm/yyyy'
@@ -51,7 +49,7 @@ export const Datepicker = (props: DatepickerProps): ReactElement => {
   }, [datetime])
 
   useClickOutside(datepickerRef, () => {
-    setOpen(false)
+    open && setOpen(false)
   })
 
   function handleDateInput(dateInput: string): void {
